Save projectVersion to notebook metadata.
This required refactor to input selects and its hooks to allow any type
of data in its state so that we can get projectVersion from selected
project.

At this stage other parts of code using this refactored code does not
complare but thats fine. It will be fixed in following commits.